Refacto Home layout
